sampling fuer mean temperature jetzt configurierbar, kompliert

+++ Brennercontrol/Globals.cpp
@@ -62,7 +62,7 @@ int max_T10 = -32768;  // new
 int min_T10 = 32767;  // new
 int max_T10_vorlauf = -32768;  // new
 int min_T10_vorlauf = 32767;  // new
-byte temp_meas_count = 0;
+int temp_meas_count = 0;
 int average_temp = 0;
 // bool send_current_T = 0;    // send once the current Temperature reading
